requirements.txt
pandas
numpy
matplotlib
scikit-learn
jupyter
streamlit
sentence-transformers
faiss-cpu
pdfplumber
python-docx
spacy
nltk
python-slugify
google-generativeai   # For Gemini API integration
requests
tqdm
# openai               # Uncomment ONLY if you want OpenAI features
pyresparser          # Uncomment if you want ATS-style parsing
seaborn